Renamed Dog Slots
