+++ Assets/Prefabs/PickupPup/Core/UI/DogCollarSlot.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: edbb79ab44434f64ea2730ad849a9318
-timeCreated: 1486154503
+guid: 1b77ef9e327a2bb4297819010cdeda24
+timeCreated: 1486654990
 licenseType: Free
 NativeFormatImporter:
   userData:
