+++ Assets/Prefabs/PickupPup/Shelter/DogShelterProfile.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 5821fc2ba76109d49b6efd1b50689b5f
-timeCreated: 1484259176
+guid: ebb1f312e95b2114e9dc0dbf8acc8e62
+timeCreated: 1486655669
 licenseType: Free
 NativeFormatImporter:
   userData:
