+++ Assets/Prefabs/PickupPup/Shelter/DogShelterSlot.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: c16f43fa6afe1284cbd64523200bb78b
-timeCreated: 1483710854
+guid: 46b99b1513974c14798ac0140b428be2
+timeCreated: 1486655510
 licenseType: Free
 NativeFormatImporter:
   userData:
