+++ Assets/Scripts/UI/PickupPup/DogSlot/DogCollarSlot.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: be58819163cad1b4894305901e643ad5
-timeCreated: 1483562845
+guid: 86ec315516dbf03458de798c835ae41f
+timeCreated: 1486655200
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
