+++ Assets/Scripts/UI/PickupPup/DogSlot/DogShelterSlot.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0350ea3b88211ce459552665aa0f239f
-timeCreated: 1483709365
+guid: 4ed0a642701c5cc48a84420cd986bc64
+timeCreated: 1486655412
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
